Conflicts:
	FutsalAgentSimulator/src/pt/ulisboa/aasma/fas/jade/agents/AgentMessages.java
	FutsalAgentSimulator/src/pt/ulisboa/aasma/fas/jade/game/Ball.java

+++ FutsalAgentSimulator/src/pt/ulisboa/aasma/fas/jade/game/Ball.java
@@ -11,11 +11,12 @@ public class Ball {
 
 	public Ball() {
 		super();
-		this.xCoord=1;
-		this.yCoord=(float) 0.5;
+		this.xCoord=1.0f;
+		this.yCoord=0.5f;
 		this.dribled = false;
 		this.passed = false;
 		this.shoted = false;
+
 	}
 	
 	public float getxCoord() {
